idiomify/requirements.txt
Markdown==3.3.4
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.1
packaging==20.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
pyparsing==2.4.7
pytorch-lightning==1.3.6
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tqdm
transformers==4.21.3
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
yarl==1.6.3
lightning-flash==0.8.0